Split out readFile from getMarkdownData
