+++ pages/blog/page/[page_index].tsx
@@ -13,7 +13,7 @@ import {
 } from '../../../components/layout'
 import { DynamicLink, BlogPagination } from '../../../components/ui'
 import { getGithubDataFromPreviewProps } from '../../../utils/github/sourceProviderConnection'
-import getMarkdownData from '../../../utils/github/getMarkdownData'
+import { getMarkdownData } from '../../../utils/getMarkdownData'
 import { useCMS } from 'tinacms'
 import OpenAuthoringSiteForm from '../../../components/layout/OpenAuthoringSiteForm'
 import { useForm } from 'tinacms'
