+++ utils/docs/getDocProps.ts
@@ -1,6 +1,6 @@
 import { getGithubDataFromPreviewProps } from '../github/sourceProviderConnection'
-import getMarkdownData from '../github/getMarkdownData'
 import getJsonData from '../github/getJsonData'
+import { getMarkdownData } from '../getMarkdownData'
 
 export async function getDocProps({ preview, previewData }: any, slug: string) {
   const {
